+++ test/w3m_account_page/w3m_account_page_ui_test.dart
@@ -9,10 +9,9 @@ import 'package:web3modal_flutter/pages/account_page.dart';
 import 'package:web3modal_flutter/services/explorer_service/explorer_service_singleton.dart';
 import 'package:web3modal_flutter/services/w3m_service/i_w3m_service.dart';
 import 'package:web3modal_flutter/utils/util.dart';
+import 'package:web3modal_flutter/web3modal_flutter.dart';
 import 'package:web3modal_flutter/widgets/web3modal_provider.dart';
 import 'package:web3modal_flutter/widgets/buttons/balance_button.dart';
-
-import 'package:walletconnect_flutter_v2/walletconnect_flutter_v2.dart';
 
 import 'package:web3modal_flutter/utils/platform/i_platform_utils.dart';
 import 'package:web3modal_flutter/utils/toast/toast_utils_singleton.dart';